+++ inject.go
@@ -23,7 +23,11 @@ package fx
 import (
 	"fmt"
 	"reflect"
+	"unicode"
+	"unicode/utf8"
 )
+
+var _typeOfIn = reflect.TypeOf(In{})
 
 // Inject fills the given struct with values from the dependency injection
 // container on application start.
@@ -42,13 +46,23 @@ func Inject(target interface{}) Option {
 	v = v.Elem()
 	t := v.Type()
 
-	// We generate a function with one argument for each field in the target
+	// We generate a function which accepts a single fx.In struct as an
+	// argument. This struct contains all exported fields of the target
 	// struct.
 
-	argTypes := make([]reflect.Type, 0, t.NumField())
+	// Fields of the generated fx.In struct.
+	fields := make([]reflect.StructField, 0, t.NumField()+1)
 
-	// List of values in the target struct aligned with the arguments of the
-	// generated function.
+	// The fix for https://github.com/golang/go/issues/18780 requires that
+	// StructField.Name is always set but versions of Go older than 1.9 expect
+	// Name to be empty for embedded fields.
+	//
+	// We use reflect_go19 and reflect_pre_go19 with build tags to implement
+	// this behavior differently in the two Go versions.
+	fields = append(fields, anonymousField(_typeOfIn))
+
+	// List of values in the target struct aligned with the fields of the
+	// generated struct.
 	//
 	// So for example, if the target is,
 	//
@@ -58,9 +72,14 @@ func Inject(target interface{}) Option {
 	// 		Baz io.Writer
 	// 	}
 	//
-	// The generated function has the shape,
+	// The generated struct has the shape,
 	//
-	// 	func(io.Reader, io.Writer)
+	// 	struct {
+	// 		fx.In
+	//
+	// 		F0 io.Reader
+	// 		F2 io.Writer
+	// 	}
 	//
 	// And `targets` is,
 	//
@@ -69,37 +88,72 @@ func Inject(target interface{}) Option {
 	// 		target.Field(2),  // Baz io.Writer
 	// 	]
 	//
-	// As we iterate through the arguments received by the function, we can
-	// simply copy the value into the corresponding value in the targets list.
+	// As we iterate through the fields of the generated struct, we can copy
+	// the value into the corresponding value in the targets list.
 	targets := make([]reflect.Value, 0, t.NumField())
 
 	for i := 0; i < t.NumField(); i++ {
 		f := t.Field(i)
-		// Skip private fields.
-		if f.PkgPath != "" {
+
+		// Skip unexported fields.
+		if f.Anonymous {
+			// If embedded, StructField.PkgPath is not a reliable indicator of
+			// whether the field is exported. See
+			// https://github.com/golang/go/issues/21122
+
+			t := f.Type
+			if t.Kind() == reflect.Ptr {
+				t = t.Elem()
+			}
+
+			if !isExported(t.Name()) {
+				continue
+			}
+		} else if f.PkgPath != "" {
 			continue
 		}
 
-		argTypes = append(argTypes, f.Type)
+		// We don't copy over names or embedded semantics.
+		fields = append(fields, reflect.StructField{
+			Name: fmt.Sprintf("F%d", i),
+			Type: f.Type,
+			Tag:  f.Tag,
+		})
 		targets = append(targets, v.Field(i))
 	}
 
 	// Equivalent to,
 	//
-	// 	func(foo Foo, bar Bar) {
-	// 		target.Foo = foo
-	// 		target.Bar = bar
+	// 	func(r struct {
+	// 		fx.In
+	//
+	// 		F1 Foo
+	// 		F2 Bar
+	// 	}) {
+	// 		target.Foo = r.F1
+	// 		target.Bar = r.F2
 	// 	}
 
 	fn := reflect.MakeFunc(
-		reflect.FuncOf(argTypes, nil /* results */, false /* variadic */),
+		reflect.FuncOf(
+			[]reflect.Type{reflect.StructOf(fields)},
+			nil,   /* results */
+			false, /* variadic */
+		),
 		func(args []reflect.Value) []reflect.Value {
-			for i, arg := range args {
-				targets[i].Set(arg)
+			result := args[0]
+			for i := 1; i < result.NumField(); i++ {
+				targets[i-1].Set(result.Field(i))
 			}
 			return nil
 		},
 	)
 
 	return Invoke(fn.Interface())
+}
+
+// isExported reports whether the identifier is exported.
+func isExported(id string) bool {
+	r, _ := utf8.DecodeRuneInString(id)
+	return unicode.IsUpper(r)
 }