+++ inject_test.go
@@ -126,22 +126,26 @@ func TestInject(t *testing.T) {
 	})
 
 	t.Run("EmbeddedUnexportedField", func(t *testing.T) {
-		var gave1 *type1
-		new1 := func() *type1 {
-			gave1 = &type1{}
-			return gave1
-		}
-
+		new1 := func() *type1 { return &type1{} }
 		var out struct{ *type1 }
 
-		app := fxtest.New(t,
-			Provide(new1),
-			Inject(&out),
-		)
-
+		app := fxtest.New(t, Provide(new1), Inject(&out))
 		defer app.MustStart().MustStop()
-		assert.NotNil(t, out.type1, "type1 must not be nil")
-		assert.True(t, gave1 == out.type1, "type1 must match")
+
+		// Unexported fields are left unchanged.
+		assert.Nil(t, out.type1, "type1 must be nil")
+	})
+
+	t.Run("EmbeddedUnexportedFieldValue", func(t *testing.T) {
+		type type4 struct{ foo string }
+		new4 := func() type4 { return type4{"foo"} }
+		var out struct{ type4 }
+
+		app := fxtest.New(t, Provide(new4), Inject(&out))
+		defer app.MustStart().MustStop()
+
+		// Unexported fields are left unchanged.
+		assert.NotEqual(t, "foo", out.type4.foo)
 	})
 
 	t.Run("DuplicateFields", func(t *testing.T) {
@@ -231,7 +235,26 @@ func TestInject(t *testing.T) {
 		assert.True(t, t2 == out.t2, "t2 must match")
 	})
 
-	t.Run("NestedDigIn", func(t *testing.T) {
+	t.Run("TopLevelDigIn", func(t *testing.T) {
+		var out struct{ dig.In }
+		app := fxtest.New(t, Inject(&out))
+		defer app.MustStart().MustStop()
+	})
+
+	t.Run("TopLevelFxIn", func(t *testing.T) {
+		new1 := func() *type1 { panic("new1 must not be called") }
+		new2 := func() *type2 { panic("new2 must not be called") }
+
+		var out struct{ In }
+		app := fxtest.New(t,
+			Provide(new1, new2),
+			Inject(&out),
+		)
+
+		defer app.MustStart().MustStop()
+	})
+
+	t.Run("NestedFxIn", func(t *testing.T) {
 		var gave1 *type1
 		new1 := func() *type1 {
 			gave1 = &type1{}
@@ -240,7 +263,7 @@ func TestInject(t *testing.T) {
 
 		var out struct {
 			Result struct {
-				dig.In
+				In
 
 				T1 *type1
 				T2 *type2 `optional:"true"`
@@ -256,6 +279,49 @@ func TestInject(t *testing.T) {
 		assert.NotNil(t, out.Result.T1, "T1 must not be nil")
 		assert.Nil(t, out.Result.T2, "T2 must be nil")
 		assert.True(t, gave1 == out.Result.T1, "T1 must match")
+	})
+
+	t.Run("FurtherNestedFxIn", func(t *testing.T) {
+		var out struct {
+			In
+
+			B struct {
+				In
+
+				C int
+			}
+		}
+
+		app := fxtest.New(t,
+			Provide(func() int { return 42 }),
+			Inject(&out),
+		)
+		defer app.MustStart().MustStop()
+		assert.Equal(t, 42, out.B.C, "B.C must match")
+	})
+
+	t.Run("FieldsCanBeOptional", func(t *testing.T) {
+		var gave1 *type1
+		new1 := func() *type1 {
+			gave1 = &type1{}
+			return gave1
+		}
+
+		var out struct {
+			T1 *type1
+			T2 *type2 `optional:"true"`
+		}
+
+		app := fxtest.New(t,
+			Provide(new1),
+			Inject(&out),
+		)
+
+		defer app.MustStart().MustStop()
+		assert.NotNil(t, out.T1, "T1 must not be nil")
+		assert.Nil(t, out.T2, "T2 must be nil")
+
+		assert.True(t, gave1 == out.T1, "T1 must match")
 	})
 }